automate-oracle-forms-gemini/automate-forms-v3/requirements.txt
pyautogui
opencv-python-headless
pytesseract
numpy
faker
pytest
allure-pytest
Flask